facein/requirements.txt
Click==7.0
dlib==19.16.0
face-recognition==1.2.3
face-recognition-models==0.3.0
imutils==0.5.2
mysql==0.0.2
mysqlclient==1.3.14
numpy==1.15.4
picamera==1.13
Pillow==5.4.1
pkg-resources==0.0.0
PyMySQL==0.9.3
pyzbar==0.1.7
RPi.GPIO==0.6.5
